Major commit:
Files query added to be used in the OpenCytoPreprocessing.r back end R script for fetching the external meta data.

OpenCytoPreprocessing.r :
Analysis description, sample group name and the concatenated study variables are now also recorded into the database.
FCS files and the XML workspace can now reside in different locations.
The association of the external meta data to the gating set has been reworked not to duplicate the joining effort: the joined meta data is now pulled from the FCSFiles table directly.
Computed hash value changed to include also the concatenated string of the study variables.

opencyto_preprocessing database schema modified:
gsTbl table now includes the gating set description, the xml workspace file path, the sample group name as well as the concatenated string of study variables.
study_vars table created to store the study variables associated with a particular analysis (gating set).

begin.html :
modified to not load the module contents if the opencyto_preprocessing schema is not present, meaning that the OpenCytoPreprocessing module has been disabled (but its interface, the web part has not been removed from the home dashboard and still rendered by the server). Once issue #18 is resolved on the server side, this would need to be removed.

ClearableComboBox.js control added to allow for a one-button approach to clearing a combo box value.

OpenCyto.css :
Streamlined by aggregating class definitions.
x-mask-loading-custom class added for a nice loading mask.
infoMask should now be on the very top layer (via the z-index tweaking).
Introduced a fix, so that buttons on a tool bar look not like text (when not hovered over), but like regular buttons.

OpenCyto.js :
Contains all of the common functionality needed for the entire suite.
Includes a customization to allow search/auto-complete in a SuperBoxCombo not just from the beginning of the word, but of any part of the word.
Also includes a custom overwrite to allow for the tab titles to be centered (for TabPanel).

OpenCytoPreprocessing.js :
FCS files and the XML workspace can now reside in different locations - necessary modifications made.
strStudyVarName modified to get populated by both the keywords as well as the external study variables.
Text fields for the analysis name as well as description added.
Alert dialog boxes changed to a generic onFailure error handler.
Layout of the main panel changed from tabs to card (i.e. 'wizard').
Grid panel with the table of files and their meta data (with a special check box column) added as another 'card' (in place of Files combo box).
Added a tool bar to the main card panel, rearranged the buttons.
Eliminated the 'Set study variables' button (card change now handles that).
Only a single empty div is left of the Ncdf functionality (at least for now).
Overall streamlined and improved error handling!